Remove Cardano.Api.Shelley.ColdKeys
